requirements.txt
# Core dependencies
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.3

# NLP and ML - Updated for Python 3.14 compatibility
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.3.0
faiss-cpu>=1.7.4
numpy>=1.26.0

# Text processing
nltk>=3.8.1
scikit-learn>=1.4.0

# Utilities
tqdm>=4.66.0
